+++ CHANGELOG.md
@@ -15,6 +15,10 @@
 * Added `near_build_info` metric which exports neard’s build information [#6680](https://github.com/near/nearcore/pull/6680)
 * Make it possible to update logging at runtime: [#6665](https://github.com/near/nearcore/pull/6665)
 * Use correct cost in gas profile for adding function call key [#6749](https://github.com/near/nearcore/pull/6749)
+* Added `path` option to `StoreConfig` which makes location to the
+  RocksDB configurable via `config.json` file (at `store.path` path)
+  rather than being hard-coded to `data` directory in neard home
+  directory.
 
 ## 1.26.0 [2022-05-18]